ABOR Unfinished

+++ PDataIO.h
@@ -18,8 +18,8 @@
 - (int) getPasvFd:(PSession*)session;
 - (int) getPortFd:(PSession*)session;
 - (BOOL) postMarkConnect:(PSession*)session;
-- (int) sendData:(PSession*)session data:(const void*)data dataSize:(unsigned long)dataSize;
-- (int) sendFile:(PSession*)session fileHandle:(NSFileHandle*)fileHandle fileSize:(unsigned long)fileSize;
-- (int) recvFile:(PSession*)session fileHandle:(NSFileHandle*)fileHandle;
+- (unsigned long long) sendData:(PSession*)session data:(const void*)data dataSize:(unsigned long long)dataSize;
+- (unsigned long long) sendFile:(PSession*)session fileHandle:(NSFileHandle*)fileHandle fileSize:(unsigned long long)fileSize;
+- (unsigned long long) recvFile:(PSession*)session fileHandle:(NSFileHandle*)fileHandle;
 
 @end
